[FAB-8599] Add makefile and build scripts.
`make evmscc-linux` produces evmscc.so in .build/linux/lib,
which is a shared library file that can be loaded into Fabric
as a system chaincode plugin.

However, there are certain limitations.
tl;dr plugin MUST be built as part of fabric

If plugin is built outside of fabric, common deps are
registered with different symbols as fabric peer binary.
For example:
evmscc.so
 \- gopath/fabric-chaincode-evm/vendor/github.com/pkgA
peer
 \- gopath/fabric/vendor/github.com/pkgA

This would result in pkgA being loaded anew when plugin is loaded,
and `init` called twice. Certain pkgs couldn't work because of the
nature of singleton design, i.e. x/net/trace registers an endpoint
at /request/Debug, and it panics if a second attempt is made.

Therefore, the build process works as following:
- make a copy of both fabric and fabric-chaincode-evm source code
  in a temporary directory on host machine
- mount them into fabric-baseimage container
- install govendor, which is then used to merge dependencies
  of fabric-chaincode-evm into fabric. In another word, only
  unique dependencies of evmscc are copied to fabric
- plugin/ is copied to fabric as source file
- statemanager/ is copied to fabric vendor
- build plugin *within* fabric

Also .gitignore is copied from Fabric.

Change-Id: Ie593fb9bdda3502484605388fdc8d373f0dfaa18
Signed-off-by: Jay Guo <guojiannan1101@gmail.com>