FAB-14092 getLogs blockhash argument implementation
 - add implementation to accept optional BlockHash argument
 - adjust comment on existing code
 - documentation of instruction
 - removed unused ledger client function

This implementation reduces the problem to an existing solved case,
that of iterating over blocks by number. This results in more requests
with the client than strictly necessary, but only one extra.

*** BlockHash Critical Fix ***

Copy the fabric 1.4 implementation of block hash calculations into our
codebase and perform fixups to all places used.

Change-Id: I176cd0eff8a892e041697720b77dbf7977230860
Signed-off-by: Morgan Bauer <mbauer@us.ibm.com>